feat: refactor commentItem structure; separate AvatarBlock, CommentBody, and CommentHeader into individual components
